生成的程序/chatbot_DefaultOrganization_20231207115651/requirements.txt
python-wechaty
